% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate-data.R
\name{add_daily_climate_data}
\alias{add_daily_climate_data}
\title{Add daily climate data to a database.}
\usage{
add_daily_climate_data(db_connection, site_number, data, type = c("rainfall",
  "max_temperature"), ...)
}
\arguments{
\item{db_connection}{The database connection to an initialised bomdata
database.}

\item{site_number}{The site number to load. Ignored if \code{site_data} is
not null.}

\item{data}{A data frame retrieved using \code{download_daily_climate_data}.
Optional. If not provided, this function will download the data.}

\item{type}{Type of data to add.}

\item{...}{Will be passed to \code{download_daily_climate_data}}
}
\description{
Add daily climate data to a database.
}
