% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hq-rainfall.R
\name{add_high_quality_rainfall}
\alias{add_high_quality_rainfall}
\title{Load the high quality daily rainfall database.}
\usage{
add_high_quality_rainfall(db_connection, tar_filename = NULL,
  url = HIGH_QUALITY_RAINFALL_URL, quiet = FALSE)
}
\arguments{
\item{db_connection}{A connection to an initialised bomdata database.}

\item{tar_filename}{Disk location of the an existing tar file}

\item{url}{Location to download tar file from, if tar_filename is NULL}

\item{quiet}{Whether to show download progress}
}
\description{
Load the high quality daily rainfall database.
}
