% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site.R
\name{bomdata_get_site_raw}
\alias{bomdata_get_site_raw}
\title{Get the site data (ie, metadata) from the BOM website}
\usage{
bomdata_get_site_raw(site_number, url = SITE_METADATA_URL)
}
\description{
Get the site data (ie, metadata) from the BOM website
}

