% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{bomdata_initialise}
\alias{bomdata_initialise}
\title{Initialise a bomdata database. The database can have been initialised in the past, so it's safe to call this again.}
\usage{
bomdata_initialise(db_connection)
}
\arguments{
\item{db_connection}{An RSQLite connection.}
}
\description{
Initialise a bomdata database. The database can have been initialised in the past, so it's safe to call this again.
}

