% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site.R
\name{bomdata_load_site}
\alias{bomdata_load_site}
\title{Retrieve and load site metadata into a database.}
\usage{
bomdata_load_site(db_connection, site_number = NULL, site_data = NULL, ...)
}
\arguments{
\item{db_connection}{The database connection to an initialised bomdata database.}

\item{site_number}{The site number to load. Ignored if \code{site_data} is not null.}

\item{site_data}{Site metadata retrieved via \code{bomdata_get_site} or \code{bomdata_get_site_raw}.}

\item{...}{Passed to \code{bomdata_get_site_raw}.}
}
\description{
Retrieve and load site metadata into a database.
}

