% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall.R
\name{download_rainfall}
\alias{download_rainfall}
\title{Get the rainfall data for a site.}
\usage{
download_rainfall(site_data, url = RAINFALL_URL)
}
\arguments{
\item{site_data}{Site metadata (as a data frame row) retrieved via
\code{get_site} or \code{download_site}}

\item{url}{The URL to retrieve the data from. The default should be correct;
read the package source for details.}
}
\description{
Get the rainfall data for a site. This is a dataframe with 4 columns, 'date',
'rainfall', 'days_measured' and quality'.
}
