% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site.R
\name{download_site}
\alias{download_site}
\title{Get the site data (ie, metadata) from the BOM website as a data.frame with a
single row.}
\usage{
download_site(site_numbers, type = c("rainfall", "max_temperature"))
}
\arguments{
\item{site_numbers}{The site number to retrieve.}

\item{type}{Which observation type to download metadata for. If the site has
no historical data for this type, returns a data frame with zero rows.}
}
\description{
Get the site data (ie, metadata) from the BOM website as a data.frame with a
single row.
}
