% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site.R
\name{get_site}
\alias{get_site}
\title{Gets the site metadata as a single-row data.frame from an existing database}
\usage{
get_site(db_connection, site_number)
}
\arguments{
\item{db_connection}{The database connection to an initialised bomdata
database.}

\item{site_number}{The site number to retrieve.}
}
\description{
Gets the site metadata as a single-row data.frame from an existing database
}

