% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site.R
\name{get_site_numbers_in_region}
\alias{get_site_numbers_in_region}
\title{Get a vector of site numbers in a specified region.}
\usage{
get_site_numbers_in_region(region_name = c("AUS", "WA", "SA", "VIC", "NSW",
  "NT", "QLD", "TAS", "ANT"), type = c("rainfall", "max_temperature"))
}
\arguments{
\item{region_name}{One of 'AUS', 'WA', 'SA', 'VIC', 'NSW', 'NT', 'QLD',
'TAS', 'ANT'.}

\item{type}{The observation for which to list sites}
}
\description{
Get a vector of site numbers in a specified region.
}
