% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise.R
\name{initialise_db}
\alias{initialise_db}
\title{Initialise a bomdata database.}
\usage{
initialise_db(db_connection)
}
\arguments{
\item{db_connection}{An RSQLite connection.}
}
\description{
Initialises a bomdata database. The database can have been initialised in
the past, so it's safe to call this again.
}
