% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{load_high_quality_rainfall}
\alias{load_high_quality_rainfall}
\title{Load the high quality daily rainfall database.}
\usage{
load_high_quality_rainfall(db_connection, tar_filename = NULL,
  url = HIGH_QUALITY_RAINFALL_URL)
}
\arguments{
\item{db_connection}{A connection to an initialised bomdata database.}

\item{tar_filename}{Disk location of the an existing tar file}

\item{url}{Location to download tar file from, if tar_filename is NULL}
}
\description{
Load the high quality daily rainfall database.
}

