% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall.R
\name{bomdata_get_rainfall_raw}
\alias{bomdata_get_rainfall_raw}
\title{Get the rainfall data for a site}
\usage{
bomdata_get_rainfall_raw(site_data, url = RAINFALL_URL)
}
\description{
Get the rainfall data for a site
}

